Add method to keep users out of categories until features are added
Learning from past mistakes and even other projects, is not wise to leave
a category blank. Some users see this and think 'oh is a bug' and then report
it on G+, Twitter, Telegram and every where they can. It can get annoying.

Adding a simple boolean helps us lock those categories and in turn have a
snackbar show up for the user that lets them know that there's no features
available at the moment. Hopefully this solves that problem and they get the
point :-)

- Removed other device extra packages. We will only have 1 to rule them all :-)

Change-Id: I5c18888d614021c29c04ee6fd53657c88fcf8ade