Simplify the structure, make changes to integrate with Settings
This commit not only simplified the structure of DU-Tweaks while keeping the
same look and feel but it makes it easier for contributors. In the old layout
the app was mainly launched from an activity. This takes us back to using Settings
preference fragment across the board which is what we used since kitkat and easier
for AOSP code.

- Removed snackbar checks

- Removed adaptive icons since is now part of Settings

- Removed OmniSwitch, Statusbar Items and Device Extra classes. We launch all three
using intents. Keeping the layouts for when we add them back.

- Simpify the check for Device Extras, using a bool

- Update bottom navigation classes, rename the names so we can better update them if
need in the future versions.

Change-Id: I867bd8e050b06d222b742e355d0d4b0c44adebad